refactor: move the /profile route into the /profile/edit route
